+++ app/src/main/java/com/example/trailquest/ui/screens/profile_screen/UserProfileScreenUiState.kt
@@ -2,8 +2,8 @@ package com.example.trailquest.ui.screens.profile_screen
 
 data class UserProfileScreenUiState(
     val userName: String = "Lau",
-    val attractionsCompleted: Float = 19f,
-    val attractionTotal: Float = 100f,
+    val attractionsCompleted: Int = 19,
+    val attractionTotal: Int = 100,
     val userTitle: String = "Newbie",
     val userLevel: Int = 1,
     val bioText: String = "I don't know",
