+++ app/src/main/java/com/example/trailquest/ui/screens/statistics_screen/StatisticsScreenUiState.kt
@@ -5,6 +5,6 @@ import com.example.trailquest.data.entities.Type
 
 
 data class StatisticsScreenUiState(
-    val statistics: HashMap<String, HashMap<Type, Array<Float>>> = hashMapOf(),
+    val statistics: HashMap<String, HashMap<Type, Array<Int>>> = hashMapOf(),
     val allCountries: List<Country> = listOf()
 )
